Nowa klasa Przekierowania, zawierająca metody dla stron jsp Krzysztof

+++ src/java/org/mypackage/hello/Dostawcaprzeniesieniedan.java
@@ -22,42 +22,45 @@ public class Dostawcaprzeniesieniedan {
             Connection conn = null;
             ResultSet rst;
             String sql,sqlw,sqll,sqld;
+            if (id == null || id.length < 1) {}
+            else {
+                try {
+                    Bazadanych baza= new Bazadanych();
+                    Class.forName(baza.getDriver2()).newInstance();
+                    conn = DriverManager.getConnection(baza.getUrl2()+baza.getDbName2(), baza.getUserName2(),baza.getPassword2());
+
+                    for (int i=0;i<id.length;i++) { ids=ids+id[i]+",";}
+
+                    ids=ids.substring(0,ids.length()-1);
+                    ids=ids+")";
+                    sql="select * from lista_dostawca where id in "+ids;
+                    Statement statement=conn.createStatement();
+                    rst=statement.executeQuery(sql);
+
+                    sqlw="insert into lista_dan values (?,?,?,?)";
+                    PreparedStatement statementw;
+                    statementw=conn.prepareStatement(sqlw);
+
+                    while (rst.next()) {            
+                            statementw.setString(1,rst.getString(1));
+                            statementw.setString(2,rst.getString(2));
+                            statementw.setString(3,rst.getString(3));
+                            statementw.setInt(4,rst.getInt(4));
+                            statementw.executeUpdate();
+                       }
+
+                    sqld="delete from lista_dostawca where id in "+ids;
+                    Statement statementd;
+                    statementd=conn.createStatement();
+                    statementd.executeUpdate(sqld);
+                }
+                catch (Exception e) { 
+                            System.out.println("Błąd połączenia z bazą danych."); 
+                            blad="Tak";
+                            }
+                finally { conn.close(); }
             
-            try {
-                Bazadanych baza= new Bazadanych();
-                Class.forName(baza.getDriver2()).newInstance();
-                conn = DriverManager.getConnection(baza.getUrl2()+baza.getDbName2(), baza.getUserName2(),baza.getPassword2());
-
-                for (int i=0;i<id.length;i++) { ids=ids+id[i]+",";}
-
-                ids=ids.substring(0,ids.length()-1);
-                ids=ids+")";
-                sql="select * from lista_dostawca where id in "+ids;
-                Statement statement=conn.createStatement();
-                rst=statement.executeQuery(sql);
-                
-                sqlw="insert into lista_dan values (?,?,?,?)";
-                PreparedStatement statementw;
-                statementw=conn.prepareStatement(sqlw);
-                               
-                while (rst.next()) {            
-                        statementw.setString(1,rst.getString(1));
-                        statementw.setString(2,rst.getString(2));
-                        statementw.setString(3,rst.getString(3));
-                        statementw.setInt(4,rst.getInt(4));
-                        statementw.executeUpdate();
-                   }
-                
-                sqld="delete from lista_dostawca where id in "+ids;
-                Statement statementd;
-                statementd=conn.createStatement();
-                statementd.executeUpdate(sqld);
             }
-            catch (Exception e) { 
-                        System.out.println("Błąd połączenia z bazą danych."); 
-                        blad="Tak";
-                        }
-            finally { conn.close(); }
     }
 
     /**
